Added check for proc texture
