+++ src/scene.cpp
@@ -165,7 +165,7 @@ int Scene::loadMaterial(string materialid) {
         Material newMaterial;
 
         //load static properties
-        for (int i = 0; i < 7; i++) {
+        for (int i = 0; i < 8; i++) {
             string line;
             utilityCore::safeGetline(fp_in, line);
             vector<string> tokens = utilityCore::tokenizeString(line);
@@ -185,6 +185,8 @@ int Scene::loadMaterial(string materialid) {
                 newMaterial.indexOfRefraction = atof(tokens[1].c_str());
             } else if (strcmp(tokens[0].c_str(), "EMITTANCE") == 0) {
                 newMaterial.emittance = atof(tokens[1].c_str());
+            }else if (strcmp(tokens[0].c_str(), "usingProcTex") == 0) {
+                newMaterial.usingProcTex = atof(tokens[1].c_str());
             }
         }
         materials.push_back(newMaterial);
